requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==3.0.8
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
pathspec==0.8.0
Pillow==7.2.0
pytz==2020.1
regex==2020.6.8
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
